requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
axblack==20220330
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3
regex==2023.8.8
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==1.26.16
